pyproject.toml
[project]
name = "finance-research"
version = "0.1.0"
description = "PhD Research in Finance"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "psycopg2-binary>=2.9.6,<3.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "statsmodels>=0.14.0",
    "yfinance>=0.2.0",
    "pandas-datareader>=0.10.0",
    "openpyxl>=3.1.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "tqdm>=4.65.0",
    "pyarrow>=14.0.1",  # Required for parquet support
    "duckdb>=0.9.0",
    "openai-batch-wrapper @ git+https://github.com/ZhengGong-hub/openai-batch-wrapper.git@v0.0.1"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 